requirements_minimal.txt
streamlit
pandas
Pillow
pytesseract
opencv-python-headless
numpy